+++ apps/app/src/test/scala/com/daml/network/integration/tests/SvStateManagementIntegrationTest2.scala
@@ -10,7 +10,7 @@ import com.daml.network.codegen.java.cn.svcrules.actionrequiringconfirmation.{
 import com.daml.network.codegen.java.cn.svcrules.coinrules_actionrequiringconfirmation.CRARC_AddFutureCoinConfigSchedule
 import com.daml.network.codegen.java.cn.svcrules.svcrules_actionrequiringconfirmation.{
   SRARC_AddMember,
-  SRARC_RemoveMember,
+  SRARC_OffboardMember,
   SRARC_SetConfig,
 }
 import com.daml.network.codegen.java.cn.svcrules.voterequestoutcome2.{
@@ -22,7 +22,7 @@ import com.daml.network.codegen.java.cn.svcrules.{
   ActionRequiringConfirmation,
   SvcRulesConfig,
   SvcRules_AddMember,
-  SvcRules_RemoveMember,
+  SvcRules_OffboardMember,
   SvcRules_SetConfig,
 }
 import com.daml.network.codegen.java.da.time.types.RelTime
@@ -40,8 +40,8 @@ import scala.jdk.OptionConverters.*
 class SvStateManagementIntegrationTest2 extends SvIntegrationTestBase {
 
   private def actionRequiring3VotesForEarlyClosing(member: String) = new ARC_SvcRules(
-    new SRARC_RemoveMember(
-      new SvcRules_RemoveMember(
+    new SRARC_OffboardMember(
+      new SvcRules_OffboardMember(
         member
       )
     )
@@ -272,8 +272,8 @@ class SvStateManagementIntegrationTest2 extends SvIntegrationTestBase {
     actAndCheck(
       "remove sv3 on svcRules contract to trigger `GarbageCollectCoinPriceVotesTrigger` to non member votes", {
         val removeAction = new ARC_SvcRules(
-          new SRARC_RemoveMember(
-            new SvcRules_RemoveMember(
+          new SRARC_OffboardMember(
+            new SvcRules_OffboardMember(
               svParties("sv3").toProtoPrimitive
             )
           )
@@ -560,7 +560,7 @@ class SvStateManagementIntegrationTest2 extends SvIntegrationTestBase {
         val sv1Party = sv1Backend.getSvcInfo().svParty
         val sv2Party = sv2Backend.getSvcInfo().svParty
         val action: ActionRequiringConfirmation = new ARC_SvcRules(
-          new SRARC_RemoveMember(new SvcRules_RemoveMember(sv1Party.toProtoPrimitive))
+          new SRARC_OffboardMember(new SvcRules_OffboardMember(sv1Party.toProtoPrimitive))
         )
 
         sv2Backend.createVoteRequest2(